(feat/refactor) Adds tuple to ARScan.swift, code cleanup.
This commit implements a tuple bestBoxSize to store the final box dimensions of the scanned object. This is easily assigned during the box volume threshold check in isReasonablySized computed property.  It may be appropriate to implement this as a computed property in its own right but the tuple works for now.